+++ CHANGELOG.md
@@ -5,13 +5,15 @@
 ### Major changes:
 ### Minor changes:
 
+	
 * The following backward compatible options to configure have been obsoleted. If you havent already migrated this code you must do this now.
   * Backend startup modes prior to 3.3.3. As enabled with `configure --with-startup-compat`. Configure option CLICON_USE_STARTUP_CONFIG is also obsoleted.
   * Configuration files (non-XML) prior to 3.3.3. As enabled with `configure --with-config-compat`. The template clicon.conf.cpp files are also removed.
   * Clixon XML C-lib prior to 3.4.0. As enabled with `configure --with-xml-compat`
   
 * new configuration option: CLICON_RESTCONF_PRETTY
-* Changed RESTCONF GET to return object referenced. ie, GET /restconf/data/X returns X. Thanks Stephen Jones for getting this right.
+* Changed restconf GET to return object referenced. ie, GET /restconf/data/X returns X. Thanks Stephen Jones for getting this right.
+* Restconf: get well-known, top-level resource, yang library version, put whole datastore, check for different keys in put lists.
 
 * Default configure file added by Matt Smith. Config file is selected in the following priority order:
   * Provide -f option when starting a program.